crm/pyproject.toml
[tool.poetry]
name = "crm"
version = "0.1.0"
description = "CRM-система в виде веб-приложения для управления клиентами."
authors = ["Belykh Igor <4340345@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.2"
django = "5.1.1"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
django-debug-toolbar = "^4.4.6"
psycopg2 = "^2.9.9"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"
factory-boy = "^3.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
